Cambio en el diseño para cargar la informacion del nivel
